--- Gopkg.toml
@@ -1,51 +1,0 @@
-[[constraint]]
-  name = "github.com/Shopify/sarama"
-  version = "~1.20.0"
-
-[[constraint]]
-  name = "github.com/linkedin/goavro"
-  version = "~2.3.0"
-
-[[constraint]]
-  name = "github.com/opentracing/opentracing-go"
-  version = "~1.0.2"
-
-[[constraint]]
-  name = "github.com/stretchr/testify"
-  version = "~1.2.1"
-
-[[constraint]]
-  name = "github.com/uber/jaeger-client-go"
-  version = "~2.14.0"
-
-[[constraint]]
-  name = "go.uber.org/zap"
-  version = "~1.9.1"
-
-[[constraint]]
-  name = "github.com/spf13/cobra"
-  version = "~0.0.3"
-
-[[constraint]]
-  name = "github.com/spf13/viper"
-  version = "~1.0.2"
-
-[[constraint]]
-  name = "github.com/spf13/pflag"
-  version = "~1.0.1"
-
-[[constraint]]
-  name = "github.com/getsentry/raven-go"
-  version = "~0.2.0"
-
-[[constraint]]
-  name = "github.com/prometheus/client_golang"
-  version = "~0.9.1"
-
-[[constraint]]
-  name = "github.com/rcrowley/go-metrics"
-  revision = "e2704e165165ec55d062f5919b4b29494e9fa790"
-
-[prune]
-  go-tests = true
-  unused-packages = true
