+++ cli/go.mod
@@ -1,0 +1,9 @@
+module github.com/spothero/tools/cli
+
+go 1.12
+
+require (
+	github.com/spf13/cobra v0.0.3
+	github.com/spf13/pflag v1.0.3
+	github.com/spf13/viper v1.3.2
+)
