+++ http/writer/go.mod
@@ -1,0 +1,5 @@
+module github.com/spothero/tools/http/writer
+
+go 1.12
+
+require github.com/gorilla/mux v1.7.1
