+++ kafka/go.mod
@@ -1,0 +1,15 @@
+module github.com/spothero/tools/kafka
+
+go 1.12
+
+require (
+	github.com/Shopify/sarama v1.22.0
+	github.com/linkedin/goavro v2.1.0+incompatible
+	github.com/opentracing/opentracing-go v1.1.0
+	github.com/prometheus/client_golang v0.9.2
+	github.com/rcrowley/go-metrics v0.0.0-20181016184325-3113b8401b8a
+	github.com/stretchr/testify v1.3.0
+	go.uber.org/atomic v1.3.2 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
+	go.uber.org/zap v1.9.1
+)
