+++ log/go.mod
@@ -1,0 +1,11 @@
+module github.com/spothero/tools/log
+
+go 1.12
+
+require (
+	github.com/prometheus/client_golang v0.9.2
+	github.com/spf13/pflag v1.0.3
+	go.uber.org/atomic v1.3.2 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
+	go.uber.org/zap v1.9.1
+)
