+++ sentry/go.mod
@@ -1,0 +1,3 @@
+module github.com/spothero/tools/sentry
+
+go 1.12
