+++ service/go.mod
@@ -1,0 +1,3 @@
+module github.com/spothero/tools/service
+
+go 1.12
