+++ time/go.mod
@@ -1,0 +1,3 @@
+module github.com/spothero/tools/time
+
+go 1.12
