+++ tracing/go.mod
@@ -1,0 +1,3 @@
+module github.com/spothero/tools/tracing
+
+go 1.12
